CCC abort scirpt if Wi-Fi interface is active
You'd want to implement it as a global "pre sanity checks" preflight
script (specified in the Advanced section of CCC's Preferences window)
in order to avoid mounting the sharepoint and destination disk image,
though (per-task preflight scripts run after CCC has verified that the
source and destination are available).

The attached script would do the trick:

Copy the file to /Library/Application Support/com.bombich.ccc/Scripts
Open CCC and click "Preferences" in the Toolbar
Click the Advanced item in the Preferences window toolbar
Click the box to select a Preflight shell script and select the
"abort_on_wifi.sh" script in that Scripts folder
Run your task with WiFi enabled and disabled to verify that it only
runs when WiFi is disabled
The script is pretty straightforward, it should work without edits. If
it doesn't, there are a couple comments in the script that you may need
to heed.